include/elf/ 	* ppc.h (R_PPC_RELAX*): Define as enum. bfd/ 	* elf32-ppc.c (ppc_elf_check_relocs): Handle R_PPC_RELAX* in switch. 	* elf32-v850.c (v850_elf_relocate_section): Warning fix.

+++ bfd/elf32-v850.c
@@ -1680,7 +1680,7 @@ v850_elf_relocate_section (bfd *output_bfd,
 		name = bfd_section_name (input_bfd, sec);
 	    }
 
-	  switch (r)
+	  switch ((int) r)
 	    {
 	    case bfd_reloc_overflow:
 	      if (! ((*info->callbacks->reloc_overflow)
